fix: separate the skills into its own app #10
